Added mini-update

+++ AutoRunManager/RegistryHelper.cs
@@ -1,19 +1,30 @@
-﻿using Microsoft.Win32;
+﻿using System.Security.AccessControl;
+using System.Security.Principal;
+using Microsoft.Win32;
 
 namespace AutoRunManager;
 
-public class RegistryHelper
+public static class RegistryHelper
 {
+    public static bool IsWindows()
+    {
+        return OperatingSystem.IsWindows();
+    }
+
     public static string[] ReadValueNames(string keyPath)
     {
+        if (!IsWindows()) return Array.Empty<string>();
+        
         using (RegistryKey key = Registry.CurrentUser.OpenSubKey(keyPath))
         {
-            return key?.GetValueNames();
+            return key?.GetValueNames() ?? Array.Empty<string>();
         }
     }
 
     public static object ReadValue(string keyPath, string valueName)
     {
+        if (!IsWindows()) return null;
+        
         using (RegistryKey key = Registry.CurrentUser.OpenSubKey(keyPath))
         {
             return key?.GetValue(valueName);
@@ -22,6 +33,8 @@ public class RegistryHelper
 
     public static void WriteValue(string keyPath, string valueName, object value)
     {
+        if (!IsWindows()) return;
+        
         using (RegistryKey key = Registry.CurrentUser.OpenSubKey(keyPath, true))
         {
             key?.SetValue(valueName, value);
@@ -30,6 +43,8 @@ public class RegistryHelper
 
     public static void DeleteValue(string keyPath, string valueName)
     {
+        if (!IsWindows()) return;
+        
         using (RegistryKey key = Registry.CurrentUser.OpenSubKey(keyPath, true))
         {
             if (key?.GetValue(valueName) != null)
@@ -41,6 +56,8 @@ public class RegistryHelper
 
     public static bool KeyExists(string keyPath)
     {
+        if (!IsWindows()) return false;
+        
         using (RegistryKey key = Registry.CurrentUser.OpenSubKey(keyPath))
         {
             return key != null;
@@ -49,9 +66,71 @@ public class RegistryHelper
 
     public static void DeleteKey(string keyPath)
     {
+        if (!IsWindows()) return;
+        
         using (RegistryKey parentKey = Registry.CurrentUser.OpenSubKey(Path.GetDirectoryName(keyPath), true))
         {
             parentKey?.DeleteSubKey(Path.GetFileName(keyPath), false);
+        }
+    }
+
+    public static string GetValueKind(string keyPath, string valueName)
+    {
+        if (!IsWindows()) return string.Empty;
+        
+        using (RegistryKey key = Registry.CurrentUser.OpenSubKey(keyPath))
+        {
+            return key?.GetValueKind(valueName).ToString() ?? string.Empty;
+        }
+    }
+
+    public static bool SetRegistryKeyAccessPermissions(string keyPath)
+    {
+        if (!IsWindows()) return false;
+
+        var user = Path.Combine(Environment.UserDomainName, Environment.UserName);
+        var registrySecurity = new RegistrySecurity();
+
+        var accessRule = new RegistryAccessRule(user,
+            RegistryRights.ReadKey | RegistryRights.WriteKey,
+            InheritanceFlags.None,
+            PropagationFlags.None,
+            AccessControlType.Allow);
+
+        registrySecurity.AddAccessRule(accessRule);
+
+        using var subKey = Registry.CurrentUser.CreateSubKey(keyPath, RegistryKeyPermissionCheck.Default, registrySecurity);
+
+        if (subKey == null) return false;
+
+        var isAdded = false;
+        var accessControl = subKey.GetAccessControl();
+        var accessRules = accessControl.GetAccessRules(true, true, typeof(NTAccount));
+
+        foreach (RegistryAccessRule rule in accessRules)
+        {
+            if (rule.IdentityReference.Value == user)
+            {
+                isAdded = true;
+                break;
+            }
+        }
+
+        return isAdded;
+    }
+
+    public static bool OpenRemoteBaseKey(string machineName)
+    {
+        if (!IsWindows()) return false;
+
+        try
+        {
+            var remoteBaseKey = RegistryKey.OpenRemoteBaseKey(RegistryHive.CurrentUser, machineName);
+            return remoteBaseKey != null;
+        }
+        catch
+        {
+            return false;
         }
     }
 }
